Write code that attempts to add on an instance of Checkout class
Note that the Checkout class hasn't even been implemented. So we will run into compilation error.
Why do this? This is actually a practice of outside-in development.
And one of the benefits is we have the perspective of the user (at the interface to our checkout class)
and anything we find we don't like, its very easy to change as we haven't actually implement yet.